requirements.txt
flask
flask_sqlalchemy
flask_login
flask_migrate
gunicorn
psycopg2-binary
werkzeug
